requirements-web.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
pytz==2023.3 